Update epics_base and epics build scripts

+++ manage/utils/files.py
@@ -1,25 +1,30 @@
 import re
-import logging as log
+import logging
 
 def match(pat, src):
-    log.debug("matching '%s': '%s':" % (src, pat))
+    logging.debug("matching '%s': '%s':" % (src, pat))
 
     with open(src, 'r') as file:
         data = file.read()
     
     return re.match(pat, flags=re.M)
 
-def substitute(rep, src, dst=None):
+def substitute(rep, src, dst=None, force=False):
     if dst is None:
         dst = src
     
-    log.debug("substituting '%s' -> '%s':" % (src, dst))
+    logging.debug("substituting '%s' -> '%s':" % (src, dst))
 
     with open(src, 'r') as file:
         data = file.read()
 
+    new_data = data
     for s, d in rep:
-        data = re.sub(s, d, data, flags=re.M)
+        new_data = re.sub(s, d, new_data, flags=re.M)
 
-    with open(dst, 'w') as file:
-        file.write(data)
+    if force or new_data != data:
+        logging.debug(f"writing file '{dst}'")
+        with open(dst, 'w') as file:
+            file.write(new_data)
+    else:
+        logging.debug(f"file unchanged '{dst}'")
